+++ ios/RCTScrollRuler/RCTScrollRulerManager.m
@@ -29,7 +29,7 @@ RCT_EXPORT_VIEW_PROPERTY(minValue, int);
 
 RCT_EXPORT_VIEW_PROPERTY(maxValue, int);
 
-RCT_EXPORT_VIEW_PROPERTY(step, int);
+RCT_EXPORT_VIEW_PROPERTY(step, float);
 
 RCT_EXPORT_VIEW_PROPERTY(defaultValue, int);
 
@@ -43,7 +43,7 @@ RCT_EXPORT_VIEW_PROPERTY(onSelect, RCTBubblingEventBlock)
 {
     
     CGFloat rullerHeight = [RCTScrollRuler rulerViewHeight];
-    _noneZeroRullerView = [[RCTScrollRuler alloc]initWithFrame:CGRectMake(10, 0, ScreenWidth-20, rullerHeight) theMinValue:0 theMaxValue:0  theStep:1 theNum:10 theUnit:@""];
+    _noneZeroRullerView = [[RCTScrollRuler alloc]initWithFrame:CGRectMake(10, 0, ScreenWidth-20, rullerHeight) theMinValue:0 theMaxValue:0  theStep:1.0 theNum:10 theUnit:@""];
     _noneZeroRullerView.bgColor = [UIColor whiteColor];
     _noneZeroRullerView.delegate        = self;
     _noneZeroRullerView.scrollByHand    = YES;
